pain: bugfix rapide
git-svn-id: https://www-lipn.univ-paris13.fr/svn/licences/trunk/outils/pain@620 04a62a05-9fe5-4f9b-8506-f2d01d604607

+++ sources/creation_pain.sql
@@ -95,7 +95,7 @@ CREATE TABLE IF NOT EXISTS pain_formation (
   nom varchar(256) COLLATE utf8_swedish_ci NOT NULL,
   annee_etude tinyint(3) unsigned NOT NULL,
   parfum varchar(40) COLLATE utf8_swedish_ci NOT NULL,
-  id_enseignant smallint(5) unsigned DEFAULT NULL,
+  id_enseignant mediumint(8) unsigned DEFAULT NULL,
   modification timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (id_formation),
   KEY id_sformation (id_sformation),
